Rename Game.js of level 1 and update imports
